requirements.txt
Django==5.0.4
djangorestframework==3.15.1
django-environ==0.11.2
psycopg2==2.9.9
django-cors-headers==4.3.1
pillow==10.3.0
pyJWT==2.1.0
social-auth-app-django==5.4.1
djoser==2.2.2
whitenoise==6.6.0
argon2-cffi==23.1.0